use long double

+++ src/analysis/calculator_analysis_cmds.cpp
@@ -22,8 +22,8 @@ using namespace utils;
 std::string classify_critical_point(
     const std::vector<std::vector<SymbolicExpression>>& hessian,
     const std::vector<std::string>& variables,
-    const std::vector<double>& values) {
-    std::vector<std::vector<double>> numeric_hessian(variables.size(), std::vector<double>(variables.size(), 0.0));
+    const std::vector<long double>& values) {
+    std::vector<std::vector<long double>> numeric_hessian(variables.size(), std::vector<long double>(variables.size(), 0.0L));
     for (std::size_t i = 0; i < variables.size(); ++i) {
         for (std::size_t j = 0; j < variables.size(); ++j) {
             SymbolicExpression current = hessian[i][j];
@@ -36,17 +36,17 @@ std::string classify_critical_point(
     }
 
     if (variables.size() == 1) {
-        double d2f = numeric_hessian[0][0];
+        long double d2f = numeric_hessian[0][0];
         if (mymath::is_near_zero(d2f, 1e-10)) return "degenerate";
-        return d2f > 0.0 ? "local min" : "local max";
+        return d2f > 0.0L ? "local min" : "local max";
     }
 
     if (variables.size() == 2) {
-        double A = numeric_hessian[0][0], B = numeric_hessian[0][1], C = numeric_hessian[1][1];
-        double D = A * C - B * B;
+        long double A = numeric_hessian[0][0], B = numeric_hessian[0][1], C = numeric_hessian[1][1];
+        long double D = A * C - B * B;
         if (mymath::is_near_zero(D, 1e-10)) return "degenerate";
-        if (D < 0.0) return "saddle point";
-        return A > 0.0 ? "local min" : "local max";
+        if (D < 0.0L) return "saddle point";
+        return A > 0.0L ? "local min" : "local max";
     }
 
     return "higher order";
@@ -60,8 +60,8 @@ bool is_infinity_literal_local(const std::string& text) {
     return value == "inf" || value == "infinity" || value == "oo";
 }
 
-std::vector<double> solve_linear_system_local(std::vector<std::vector<double>> matrix,
-                                              std::vector<double> rhs) {
+std::vector<long double> solve_linear_system_local(std::vector<std::vector<long double>> matrix,
+                                              std::vector<long double> rhs) {
     const std::size_t n = rhs.size();
     for (std::size_t col = 0; col < n; ++col) {
         std::size_t pivot = col;
@@ -77,12 +77,12 @@ std::vector<double> solve_linear_system_local(std::vector<std::vector<double>> m
             std::swap(matrix[pivot], matrix[col]);
             std::swap(rhs[pivot], rhs[col]);
         }
-        const double divisor = matrix[col][col];
+        const long double divisor = matrix[col][col];
         for (std::size_t c = col; c < n; ++c) matrix[col][c] /= divisor;
         rhs[col] /= divisor;
         for (std::size_t row = 0; row < n; ++row) {
             if (row == col) continue;
-            const double factor = matrix[row][col];
+            const long double factor = matrix[row][col];
             if (mymath::is_near_zero(factor, 1e-14)) continue;
             for (std::size_t c = col; c < n; ++c) matrix[row][c] -= factor * matrix[col][c];
             rhs[row] -= factor * rhs[col];
@@ -112,7 +112,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
             analysis = ctx.build_analysis(arguments[0]); point_arg = arguments[1]; dir_idx = 2;
         }
         int dir = 0; if (arguments.size() > dir_idx) dir = static_cast<int>(round_to_long_long(ctx.parse_decimal(arguments[dir_idx])));
-        double limit_value = 0.0;
+        long double limit_value = 0.0L;
         try {
             limit_value = ctx.normalize_result(analysis.limit(ctx.parse_decimal(point_arg), dir));
         } catch (const std::runtime_error& ex) {
@@ -158,21 +158,21 @@ bool handle_analysis_command(const AnalysisContext& ctx,
         }
 
         // Find critical points by solving gradient = 0
-        std::vector<std::map<std::string, double>> critical_points;
+        std::vector<std::map<std::string, long double>> critical_points;
 
         if (variables.size() == 1) {
             const std::string& variable = variables[0];
             const SymbolicExpression derivative = gradient[0];
-            auto eval_derivative = [&](double x) {
+            auto eval_derivative = [&](long double x) {
                 SymbolicExpression at_x =
                     derivative.substitute(variable, SymbolicExpression::number(x)).simplify();
-                double value = 0.0;
+                long double value = 0.0L;
                 if (!at_x.is_number(&value)) {
                     throw std::runtime_error("derivative is not numeric at this point");
                 }
                 return value;
             };
-            auto add_point = [&](double x) {
+            auto add_point = [&](long double x) {
                 for (const auto& existing : critical_points) {
                     const auto it = existing.find(variable);
                     if (it != existing.end() && mymath::abs(it->second - x) < 1e-5) {
@@ -184,39 +184,39 @@ bool handle_analysis_command(const AnalysisContext& ctx,
 
             // Extended scan range: [-100, 100] with adaptive refinement
             // Also scan for even-multiplicity roots by checking second derivative sign changes
-            const double scan_min = -100.0;
-            const double scan_max = 100.0;
+            const long double scan_min = -100.0L;
+            const long double scan_max = 100.0L;
             const int coarse_segments = 512;
 
             // Pre-compute second derivative for validation
             SymbolicExpression second_deriv = derivative.derivative(variable).simplify();
-            auto eval_second = [&](double x) {
+            auto eval_second = [&](long double x) {
                 SymbolicExpression at_x = second_deriv.substitute(variable, SymbolicExpression::number(x)).simplify();
-                double value = 0.0;
+                long double value = 0.0L;
                 if (at_x.is_number(&value)) return value;
-                return 0.0;
+                return 0.0L;
             };
 
-            double previous_x = scan_min;
-            double previous_value = eval_derivative(previous_x);
+            long double previous_x = scan_min;
+            long double previous_value = eval_derivative(previous_x);
             for (int i = 1; i <= coarse_segments; ++i) {
-                const double current_x = scan_min + (scan_max - scan_min) * static_cast<double>(i) / coarse_segments;
-                const double current_value = eval_derivative(current_x);
+                const long double current_x = scan_min + (scan_max - scan_min) * static_cast<long double>(i) / coarse_segments;
+                const long double current_value = eval_derivative(current_x);
                 // Sign change indicates a root of f' = 0
-                if ((previous_value < 0.0 && current_value > 0.0) ||
-                    (previous_value > 0.0 && current_value < 0.0)) {
-                    double left = previous_x;
-                    double right = current_x;
-                    double left_value = previous_value;
+                if ((previous_value < 0.0L && current_value > 0.0L) ||
+                    (previous_value > 0.0L && current_value < 0.0L)) {
+                    long double left = previous_x;
+                    long double right = current_x;
+                    long double left_value = previous_value;
                     for (int iter = 0; iter < 80; ++iter) {
-                        const double mid = (left + right) * 0.5;
-                        const double mid_value = eval_derivative(mid);
+                        const long double mid = (left + right) * 0.5;
+                        const long double mid_value = eval_derivative(mid);
                         if (mymath::is_near_zero(mid_value, 1e-12)) {
                             left = right = mid;
                             break;
                         }
-                        if ((left_value < 0.0 && mid_value > 0.0) ||
-                            (left_value > 0.0 && mid_value < 0.0)) {
+                        if ((left_value < 0.0L && mid_value > 0.0L) ||
+                            (left_value > 0.0L && mid_value < 0.0L)) {
                             right = mid;
                         } else {
                             left = mid;
@@ -234,9 +234,9 @@ bool handle_analysis_command(const AnalysisContext& ctx,
             // This distinguishes true even-multiplicity roots from flat regions where both f' and f'' approach zero
 
             for (int i = 0; i <= coarse_segments; ++i) {
-                const double x = scan_min + (scan_max - scan_min) * static_cast<double>(i) / coarse_segments;
-                const double deriv_val = eval_derivative(x);
-                const double second_val = eval_second(x);
+                const long double x = scan_min + (scan_max - scan_min) * static_cast<long double>(i) / coarse_segments;
+                const long double deriv_val = eval_derivative(x);
+                const long double second_val = eval_second(x);
 
                 // For even-multiplicity roots: f' ≈ 0 but f'' ≠ 0
                 // For flat regions: both f' ≈ 0 and f'' ≈ 0
@@ -244,10 +244,10 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                 if (mymath::abs(deriv_val) < 1e-8 && mymath::abs(second_val) > 1e-6) {
                     // This is likely an even-multiplicity root (derivative touches zero)
                     // Refine using Newton's method on f'
-                    double refined_x = x;
+                    long double refined_x = x;
                     for (int iter = 0; iter < 20; ++iter) {
-                        const double f_prime = eval_derivative(refined_x);
-                        const double f_double_prime = eval_second(refined_x);
+                        const long double f_prime = eval_derivative(refined_x);
+                        const long double f_double_prime = eval_second(refined_x);
                         if (mymath::is_near_zero(f_prime, 1e-12)) break;
                         if (mymath::abs(f_double_prime) < 1e-8) break;
                         refined_x = refined_x - f_prime / f_double_prime;
@@ -261,28 +261,28 @@ bool handle_analysis_command(const AnalysisContext& ctx,
             }
         } else {
             auto eval_gradient_at = [&](const SymbolicExpression& g,
-                                        const std::map<std::string, double>& point) {
+                                        const std::map<std::string, long double>& point) {
                 SymbolicExpression current = g;
                 for (const auto& [name, value] : point) {
                     current = current.substitute(name, SymbolicExpression::number(value)).simplify();
                 }
-                double numeric = 0.0;
+                long double numeric = 0.0L;
                 if (!current.is_number(&numeric)) {
                     throw std::runtime_error("gradient is not numeric at this point");
                 }
                 return numeric;
             };
 
-            auto gradient_norm_at = [&](const std::map<std::string, double>& point) {
-                double norm = 0.0;
+            auto gradient_norm_at = [&](const std::map<std::string, long double>& point) {
+                long double norm = 0.0L;
                 for (const auto& g : gradient) {
-                    const double val = eval_gradient_at(g, point);
+                    const long double val = eval_gradient_at(g, point);
                     norm += val * val;
                 }
                 return mymath::sqrt(norm);
             };
 
-            auto add_critical_point = [&](const std::map<std::string, double>& point) {
+            auto add_critical_point = [&](const std::map<std::string, long double>& point) {
                 bool duplicate = false;
                 for (const auto& existing : critical_points) {
                     bool same = true;
@@ -298,7 +298,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                     if (same) { duplicate = true; break; }
                 }
                 if (!duplicate) {
-                    std::map<std::string, double> normalized;
+                    std::map<std::string, long double> normalized;
                     for (const auto& [k, v] : point) {
                         normalized[k] = ctx.normalize_result(v);
                     }
@@ -307,19 +307,19 @@ bool handle_analysis_command(const AnalysisContext& ctx,
             };
 
             // Try Newton-Raphson from multiple starting points
-            std::vector<std::map<std::string, double>> starting_points;
+            std::vector<std::map<std::string, long double>> starting_points;
 
             // Origin as starting point
-            std::map<std::string, double> origin;
-            for (const auto& v : variables) origin[v] = 0.0;
+            std::map<std::string, long double> origin;
+            for (const auto& v : variables) origin[v] = 0.0L;
             starting_points.push_back(origin);
 
             // Grid of starting points for better coverage
-            const std::vector<double> grid_values = {-10.0, -5.0, 0.0, 5.0, 10.0};
+            const std::vector<long double> grid_values = {-10.0L, -5.0, 0.0L, 5.0, 10.0L};
             if (variables.size() == 2) {
-                for (double v0 : grid_values) {
-                    for (double v1 : grid_values) {
-                        std::map<std::string, double> pt;
+                for (long double v0 : grid_values) {
+                    for (long double v1 : grid_values) {
+                        std::map<std::string, long double> pt;
                         pt[variables[0]] = v0;
                         pt[variables[1]] = v1;
                         starting_points.push_back(pt);
@@ -329,7 +329,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
 
             for (const auto& start : starting_points) {
                 try {
-                    std::map<std::string, double> current = start;
+                    std::map<std::string, long double> current = start;
                     if (gradient_norm_at(current) < 1e-8) {
                         add_critical_point(current);
                         continue;
@@ -338,23 +338,23 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                     // Newton-Raphson iteration
                     for (int iter = 0; iter < 50; ++iter) {
                         // Build Jacobian of gradient (Hessian of original function)
-                        std::vector<double> rhs(variables.size(), 0.0);
-                        std::vector<std::vector<double>> jac(variables.size(),
-                            std::vector<double>(variables.size(), 0.0));
+                        std::vector<long double> rhs(variables.size(), 0.0L);
+                        std::vector<std::vector<long double>> jac(variables.size(),
+                            std::vector<long double>(variables.size(), 0.0L));
 
                         for (std::size_t row = 0; row < variables.size(); ++row) {
                             rhs[row] = -eval_gradient_at(gradient[row], current);
                             for (std::size_t col = 0; col < variables.size(); ++col) {
-                                std::map<std::string, double> perturbed = current;
+                                std::map<std::string, long double> perturbed = current;
                                 perturbed[variables[col]] += 1e-6;
                                 jac[row][col] = (eval_gradient_at(gradient[row], perturbed) -
                                                 eval_gradient_at(gradient[row], current)) / 1e-6;
                             }
                         }
 
-                        const std::vector<double> delta = solve_linear_system_local(jac, rhs);
+                        const std::vector<long double> delta = solve_linear_system_local(jac, rhs);
 
-                        double max_change = 0.0;
+                        long double max_change = 0.0L;
                         for (std::size_t i = 0; i < variables.size(); ++i) {
                             current[variables[i]] += delta[i];
                             max_change = std::max(max_change, mymath::abs(delta[i]));
@@ -364,7 +364,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                     }
 
                     // Check if this is a valid critical point
-                    const double grad_norm = gradient_norm_at(current);
+                    const long double grad_norm = gradient_norm_at(current);
                     if (grad_norm < 1e-8) {
                         add_critical_point(current);
                     }
@@ -406,7 +406,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                 for (const auto& [v, val] : pt) {
                     second_at_pt = second_at_pt.substitute(v, SymbolicExpression::number(val)).simplify();
                 }
-                double second_val = 0.0;
+                long double second_val = 0.0L;
                 second_at_pt.is_number(&second_val);
                 if (second_val > 1e-10) out << " (local min)";
                 else if (second_val < -1e-10) out << " (local max)";
@@ -416,7 +416,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                 auto hessian = expr.hessian(variables);
 
                 // Evaluate Hessian at critical point
-                std::vector<std::vector<double>> hessian_values(hessian.size(), std::vector<double>(hessian.size(), 0.0));
+                std::vector<std::vector<long double>> hessian_values(hessian.size(), std::vector<long double>(hessian.size(), 0.0L));
                 bool hessian_evaluable = true;
                 for (size_t r = 0; r < hessian.size(); ++r) {
                     for (size_t c = 0; c < hessian[r].size(); ++c) {
@@ -445,7 +445,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
 
                     // For 2x2, also check determinant
                     if (hessian_values.size() == 2) {
-                        double det = hessian_values[0][0] * hessian_values[1][1] - hessian_values[0][1] * hessian_values[1][0];
+                        long double det = hessian_values[0][0] * hessian_values[1][1] - hessian_values[0][1] * hessian_values[1][0];
                         if (det <= 1e-10) positive_definite = false;
                         if (det <= 1e-10) negative_definite = false;
                     }
@@ -458,7 +458,7 @@ bool handle_analysis_command(const AnalysisContext& ctx,
                         // Check if Hessian is zero (degenerate case)
                         bool all_zero = true;
                         for (const auto& row : hessian_values) {
-                            for (double val : row) {
+                            for (long double val : row) {
                                 if (mymath::abs(val) > 1e-10) {
                                     all_zero = false;
                                     break;
